Fix mate sequence when reading sam/bam
Previously upon reading the case of tid == mtid was detected and the
sequence name mapped to "=". This causes missing sequence name errors
upon decompression. As the case of tid == mtid is handled during writing
of sam/bam, this patch simply records the full mate sequence name,
resolving the matching issues.